+++ src/admin_commands/viewdefaultfilteredwordslist.js
@@ -1,6 +1,6 @@
 const databasefunctions = require('../functions/databasefunctions.js');
 const messagefunctions = require('../functions/messages.js');
-const constants = require('../constants/constants.js');
+const constants = require('../assets/config.js');
 
 module.exports = {
     name: 'viewdefaultfilteredwordslist',