+++ src/general_commands/specs.js
@@ -1,6 +1,6 @@
 const messages = require('../functions/messages.js');
 const databasefunctions = require('../functions/databasefunctions.js');
-const constants = require('../constants/constants.js');
+const constants = require('../assets/config.js');
 const { EmbedBuilder } = require('discord.js');
 
 module.exports = {