+++ src/general_commands/suggestion.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 const databasefunctions = require('../functions/databasefunctions.js');
-const constants = require('../constants/constants.js');
+const constants = require('../assets/config.js');
 const logging = require('../functions/logging.js');
 
 module.exports = {